LCA discrepancy test
Code : why did the LCA with the two dataset gave such different results. Variables d_... are exactly the same on the two datasets. The changes seem to be in the origin_tous_g2(bis?) and d_racism/d_racpot variables. 
Finally, the filtered dataset are similar and the results of the the 2 LCAs are similar 2 but not identical. I do not know why we don't have such a discrepancy as before nor why there is still a little bit of discrepancy...